+++ packages/block-library/src/search/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { __ } from '@wordpress/i18n';
+import { __, _x } from '@wordpress/i18n';
 import { search as icon } from '@wordpress/icons';
 
 /**
@@ -16,7 +16,7 @@ const { name } = metadata;
 export { metadata, name };
 
 export const settings = {
-	title: __( 'Search' ),
+	title: _x( 'Search', 'block title' ),
 	description: __( 'Help visitors find your content.' ),
 	icon,
 	keywords: [ __( 'find' ) ],